i fix the unused interface(Implement the interfaces in the missing classes)

+++ ecommerce/src/main/java/v/o/r/ecommerce/common/interfaces/products/IProducts.java
@@ -2,8 +2,8 @@ package v.o.r.ecommerce.common.interfaces.products;
 
 public interface IProducts {
 
-    int getId();
-    void setId(int id);
+    Long getId();
+    void setId(Long id);
     String getName();
     void setName(String name);
     String getDescription();
